backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
python-jose==3.4.0
passlib==1.7.4
requests==2.32.3
python-multipart==0.0.20

# --- LangChain Ecosystem ---
langchain==0.3.24
langchain-core==0.3.66
langchain-google-genai==2.1.4
langchain-openai==0.3.23
langchain-pinecone==0.2.5
langchain-text-splitters==0.3.8

# --- Pinecone + Llama ---
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index-core==0.12.31
llama-index-readers-file==0.4.7

# --- Other deps ---
groq==0.22.0
pandas==2.2.3
starlette==0.45.3
typing_extensions==4.12.2
bcrypt==4.2.1
aiohttp==3.10.11
async-timeout==4.0.3
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
sniffio==1.3.1
anyio==4.8.0
httpcore==1.0.7
httpx==0.28.1
jiter==0.8.2
psycopg2-binary==2.9.9
protobuf==4.25.3
googleapis-common-protos==1.62.0
pymilvus==2.5.3
openai==1.84.0
tensorflow==2.16.2
